Выполнил задание #3. Учебный проект: время разбивать камни
